test/fixtures/wpt/performance-timeline/navigation-id-reset.tentative.html

<!DOCTYPE HTML>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/html/browsers/browsing-the-web/back-forward-cache/resources/helper.sub.js"></script>
<script>
  const reload = () => {
    window.location.reload();
  };

  const getNavigationId = () => {
    window.performance.mark('initial_load');
    let entries = window.performance.getEntriesByType('mark');
    return entries[entries.length - 1].navigationId;
  }

  promise_test(async t => {
    const pageA = new RemoteContext(token());
    const pageB = new RemoteContext(token());

    const urlA = executorPath + pageA.context_id;
    const urlB = originCrossSite + executorPath + pageB.context_id;
    // Open url A.
    window.open(urlA, '_blank', 'noopener')
    await pageA.execute_script(waitForPageShow);

    let navigationIdInitial = await pageA.execute_script(getNavigationId);

    // Navigate away to url B and back.
    await navigateAndThenBack(pageA, pageB, urlB);

    // Assert navigation id is re-generated and thus different when the
    // document is load from bfcache.
    navigationIdAfterBFCacheNav = await pageA.execute_script(getNavigationId);
    assert_not_equals(navigationIdInitial, navigationIdAfterBFCacheNav, 'Navigation Id should be \
    re-generated and different from the previous one after back-forward-cache navigation.');

    // Reload page.
    await pageA.execute_script(reload);
    await pageA.execute_script(waitForPageShow);

    navigationIdAfterReset = await pageA.execute_script(getNavigationId);

    assert_not_equals(navigationIdAfterReset, navigationIdAfterBFCacheNav, 'Navigation Id should\
    be re-generated after reload which is different from the previous one.');

    assert_not_equals(navigationIdAfterReset, navigationIdInitial, 'Navigation Id should\
    be re-generated after reload which is different from the one of the initial load.');

  }, 'Navigation Id should be re-generated after reload.');
</script>
